Moved retryOperation back to core
Closes #19
